--- app/src/main/java/com/rodcollab/lingoleap/collections/history/repository/SearchHistory.kt
@@ -1,11 +1,0 @@
-package com.rodcollab.lingoleap.collections.history.repository
-
-import com.rodcollab.lingoleap.features.word.search.SearchedWordDomain
-import com.rodcollab.lingoleap.features.word.search.Word
-
-interface SearchHistory {
-
-    suspend fun add(word: Word)
-
-    suspend fun getList(): List<SearchedWordDomain>
-}
