--- app/src/main/java/com/rodcollab/lingoleap/collections/saved/domain/WordsSavedUseCase.kt
@@ -1,8 +1,0 @@
-package com.rodcollab.lingoleap.collections.saved.domain
-
-import com.rodcollab.lingoleap.features.profile.SavedWordItemState
-
-interface WordsSavedUseCase {
-
-    suspend operator fun invoke(): List<SavedWordItemState>
-}
