--- app/src/main/java/com/rodcollab/lingoleap/collections/saved/repository/WordsSavedRepository.kt
@@ -1,13 +1,0 @@
-package com.rodcollab.lingoleap.collections.saved.repository
-
-import com.rodcollab.lingoleap.features.word.search.WordSaved
-
-interface WordsSavedRepository {
-
-    suspend fun unsavedWord(name:String)
-
-    suspend fun savedWord(name: String)
-
-    suspend fun getSavedWords(): List<WordSaved>
-
-}
