--- app/src/main/java/com/rodcollab/lingoleap/collections/search/domain/GetWordUseCase.kt
@@ -1,8 +1,0 @@
-package com.rodcollab.lingoleap.collections.search.domain
-
-import com.rodcollab.lingoleap.features.word.search.Word
-
-interface GetWordUseCase {
-    suspend operator fun invoke(query: String) : List<Word>
-
-}
