--- app/src/main/java/com/rodcollab/lingoleap/collections/search/domain/SaveWord.kt
@@ -1,7 +1,0 @@
-package com.rodcollab.lingoleap.collections.search.domain
-
-interface SaveWord {
-
-    suspend operator fun invoke(name:String)
-
-}
