+++ app/src/main/java/com/rodcollab/lingoleap/features/history/repository/SearchHistory.kt
@@ -1,0 +1,11 @@
+package com.rodcollab.lingoleap.features.history.repository
+
+import com.rodcollab.lingoleap.features.search.SearchedWordDomain
+import com.rodcollab.lingoleap.features.search.Word
+
+interface SearchHistory {
+
+    suspend fun add(word: Word)
+
+    suspend fun getList(): List<SearchedWordDomain>
+}
