+++ app/src/main/java/com/rodcollab/lingoleap/features/search/SearchViewModel.kt
@@ -1,12 +1,9 @@
-package com.rodcollab.lingoleap.features.word.search
+package com.rodcollab.lingoleap.features.search
 
-import android.util.Log
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.rodcollab.lingoleap.collections.history.repository.SearchHistory
-import com.rodcollab.lingoleap.collections.saved.repository.WordsSavedRepository
-import com.rodcollab.lingoleap.collections.search.domain.GetWordUseCase
-import com.rodcollab.lingoleap.collections.search.domain.SaveWord
+import com.rodcollab.lingoleap.features.history.repository.SearchHistory
+import com.rodcollab.lingoleap.features.search.domain.GetWordUseCase
 import com.rodcollab.lingoleap.features.word.WordItemUiState
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.Dispatchers
@@ -19,8 +16,6 @@ import javax.inject.Inject
 @HiltViewModel
 class SearchViewModel @Inject constructor(
     private val searchHistory: SearchHistory,
-    private val wordsSavedRepository: WordsSavedRepository,
-    private val saveWord: SaveWord,
     private val getWord: GetWordUseCase
 ) : ViewModel() {
 
@@ -96,14 +91,7 @@ class SearchViewModel @Inject constructor(
             }
             is SearchEvent.OnSaveWord -> {
                 viewModelScope.launch {
-                    saveWord(event.word.word)
-                    Log.d("savedWords", wordsSavedRepository.getSavedWords().toString())
-
                     _state.value.infoItem.let { itemClicked ->
-                        val saved = withContext(Dispatchers.IO) {
-                            wordsSavedRepository.getSavedWords()
-                                .any { itemClicked.word == it.name }
-                        }
                         _state.value = SearchState(
                             query = _state.value.query,
                             isHintVisible = _state.value.isHintVisible,
@@ -114,7 +102,6 @@ class SearchViewModel @Inject constructor(
                                 word = itemClicked.word,
                                 meanings = itemClicked.meanings,
                                 audio = itemClicked.audio,
-                                saved = saved
                             )
                         )
                     }
@@ -164,9 +151,7 @@ class SearchViewModel @Inject constructor(
                     infoItem = InfoItemClicked(
                         word = itemClicked.word,
                         meanings = itemClicked.meanings,
-                        audio = itemClicked.audio,
-                        saved = wordsSavedRepository.getSavedWords()
-                            .any { itemClicked.word == it.name })
+                        audio = itemClicked.audio,)
                 )
             }
         }
