+++ app/src/main/java/com/rodcollab/lingoleap/features/search/domain/GetWordUseCase.kt
@@ -1,0 +1,8 @@
+package com.rodcollab.lingoleap.features.search.domain
+
+import com.rodcollab.lingoleap.features.search.Word
+
+interface GetWordUseCase {
+    suspend operator fun invoke(query: String) : List<Word>
+
+}
